+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes
 
 ### Unreleased
 
+* 2022-02-03 - Improvement: Silently skip the guest user in the scheduled task
 * 2022-02-03 - Bugfix: Change default dayofweek of scheduled task from 0 to *
 
 ### v3.9-r1